Making the progressbar effective
